PayMethods with validations finished

+++ Sprint2/src/controllers/payMethods.controllers.js
@@ -7,29 +7,54 @@ exports.getAll = async (req, res) => {
 };
 
 exports.create = async (req, res) => {
-    const { payMeth } = req.body;
-    const newPayMeth = await db.PayMethods.create({ 
-        payMeth
+    const { NamePay } = req.body;
+    const  FindPayName  = await db.PayMethods.findOne({
+        where: {
+            NamePay 
+        }
     });
-    res.json(newPayMeth);
+    
+    if( NamePay == FindPayName.NamePay ){
+        res.status(400).json('El nombre de metodo de pago ya existe, porfavor verifique su solicitud');
+    }else{
+        const newPayMeth = await db.PayMethods.create({ NamePay });
+        res.status(200).json(newPayMeth);
+    }
 };
 
 exports.update = async (req, res) => {
     const { id } = req.params;
-    const { payMeth } = req.body;
-    const updatePayMeth = await db.PayMethods.update({ payMeth },{
-        where:{
-            id
-        }
+    const { NamePay } = req.body;
+    const IdPayMeth = await db.PayMethods.findOne({
+        where: { id }
     });
-    res.json('Medio de Pago Actualizado');
+    
+    if( IdPayMeth ){
+        const updatePayMeth = await db.PayMethods.update({ NamePay },{
+            where:{
+                id
+            }
+        });
+        res.status(201).json(updatePayMeth);
+    }else{
+        res.status(400).json('El id del metodo de pago no existe, porfavor verifique su solicitud');
+    }
 };
 
 exports.destroy = async (req, res) => {
     const { id } = req.params;
-    const deletePayMeth = await db.PayMethods.destroy({
-        where: {
-            id: id 
-        }
+    const IdPayMeth = await db.PayMethods.findOne({
+        where: { id }
     });
+
+    if(IdPayMeth){
+        await db.PayMethods.destroy({
+            where: {
+                id: id 
+            }
+        });
+        res.status(200).json('Metodo de pago aliminado');
+    }else{
+        res.status(400).json('El id del metodo de pago no existe, porfavor verifique su solicitud');      
+    }
 };
